chore(format): apply Biome formatting on consolidation branch
