+++ src/app/admin/dashboard/page.tsx
@@ -18,7 +18,7 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Skeleton } from '@/components/ui/skeleton';
 import { useAdmin } from '@/hooks/use-admin';
-import { registrationService } from '@/lib/firebase-admin';
+import { registrationService } from '@/lib/services';
 
 interface DashboardStats {
   registrations: {
@@ -91,7 +91,7 @@ function QuickActionCard({
 
   if (href) {
     return (
-      <Link href={href}>
+      <Link href={href as any}>
         <Card className="hover:bg-accent h-full cursor-pointer transition-colors">
           <CardContent className="pt-6">{content}</CardContent>
         </Card>
