+++ src/lib/services/admin-service.ts
@@ -162,4 +162,3 @@ export const adminService = {
 };
 
 export default adminService;
-
