+++ src/lib/services/announcement-service.ts
@@ -123,4 +123,3 @@ export const announcementService = {
 };
 
 export default announcementService;
-
