+++ src/lib/services/calendar-service.ts
@@ -262,4 +262,3 @@ export const calendarService = {
 };
 
 export default calendarService;
-
