+++ src/lib/services/index.ts
@@ -20,4 +20,3 @@ export {
   type RegistrationWithDetails,
   registrationService,
 } from './registration-service';
-
